+++ appsiel/app/Nomina/ModosLiquidacion/PrestacionesSociales/Vacaciones.php
@@ -162,7 +162,7 @@ class Vacaciones implements Estrategia
             $programacion_vacaciones->save();
 
             // Almacenar registro para los días no hábiles
-            $this->crear_registro_concepto_vacaciones_dias_no_habiles( $programacion_vacaciones->id, $liquidacion['documento_nomina'], $liquidacion['empleado'], $libro_vacaciones->dias_no_habiles * (int)config('nomina.horas_dia_laboral') );
+            $this->crear_registro_concepto_vacaciones_dias_no_habiles( $programacion_vacaciones->id, $liquidacion['documento_nomina'], $liquidacion['empleado'], $libro_vacaciones->dias_no_habiles * (float)config('nomina.horas_dia_laboral') );
 
             $this->actualizar_libro_vacaciones( $liquidacion['empleado'], $liquidacion['documento_nomina'], $libro_vacaciones );
         }        
@@ -171,7 +171,7 @@ class Vacaciones implements Estrategia
 
         return [
                     [
-                        'cantidad_horas' => $cantidad_dias_amortizar * (int)config('nomina.horas_dia_laboral'), // Se almacenan solo los días a amortizar, el resto del tiempo de vacaciones se almacenan en el documento siguiente de nómina
+                        'cantidad_horas' => $cantidad_dias_amortizar * (float)config('nomina.horas_dia_laboral'), // Se almacenan solo los días a amortizar, el resto del tiempo de vacaciones se almacenan en el documento siguiente de nómina
                         'valor_devengo' => $valores->devengo,
                         'valor_deduccion' => $valores->deduccion,
                         'novedad_tnl_id' => $this->novedad_id,
@@ -228,7 +228,7 @@ class Vacaciones implements Estrategia
 
         $this->tabla_resumen['base_liquidacion'] = $parametros_prestacion->base_liquidacion;
 
-        $this->tabla_resumen['cantidad_dias_salario'] = (int)config('nomina.horas_laborales') / (int)config('nomina.horas_dia_laboral');
+        $this->tabla_resumen['cantidad_dias_salario'] = (float)config('nomina.horas_laborales') / (float)config('nomina.horas_dia_laboral');
 
         switch ( $parametros_prestacion->base_liquidacion )
         {
